Fix builds named 'Empty' using empty style
Renamed `ShowPreview` to `IsEmpty` as it is pretty much used in the same
way.
Updated the data trigger used for the empty style changes to look at
`IsEmpty`, rather than `Name`. We also now set the text to 'Empty' in the XAML and no longer change the `Name` property on the model.