Enhancement/style changes (#10)
* Makes SpecializationSlot Trait properties array based

Making them be backed by an array allows for easy iteration over the slots.

* Renames brushes

Renames `FilledTextForegoundBrush` -> `defaultForegoundBrush`.
Adds `defaultForegroundPen` and `darkerForegoundBrush`.

* Updates SlotTraitChoices to use static brushes

Now the pen and brush used to draw are passed into the control.
Updates the render logic to reduce code duplication.

* Reflows SlotTraitChoices property definitions

* Define thickness on defaultForegroundPen

* Fix builds named 'Empty' using empty style

Renamed `ShowPreview` to `IsEmpty` as it is pretty much used in the same
way.
Updated the data trigger used for the empty style changes to look at
`IsEmpty`, rather than `Name`. We also now set the text to 'Empty' in the XAML and no longer change the `Name` property on the model.

* Removes unused xml namespaces from MainWindow.xaml

* Adds text trimming to the build names

This stops name getting cut-off by the rename and delete buttons and
looking odd.